[nxp][platform][k32w0] Update files after moving some components
