+++ src/platform/nxp/k32w0/BUILD.gn
@@ -34,8 +34,8 @@ static_library("nxp_platform") {
   defines = []
   sources = [
     "../../SingletonConfigurationManager.cpp",
-    "../common/legacy/BLEManagerCommon.cpp",
-    "../common/legacy/BLEManagerCommon.h",
+    "../common/ble/BLEManagerCommon.cpp",
+    "../common/ble/BLEManagerCommon.h",
     "BLEManagerImpl.cpp",
     "BLEManagerImpl.h",
     "CHIPDevicePlatformConfig.h",
@@ -83,13 +83,13 @@ static_library("nxp_platform") {
   }
 
   if (chip_enable_ota_requestor) {
-    public += [ "../common/legacy/OTAImageProcessorImpl.h" ]
+    public += [ "../common/ota/OTAImageProcessorImpl.h" ]
 
     sources += [
-      "../common/legacy/OTAImageProcessorImpl.cpp",
-      "../common/legacy/OTAImageProcessorImpl.h",
-      "../common/legacy/OTATlvProcessor.cpp",
-      "../common/legacy/OTATlvProcessor.h",
+      "../common/ota/OTAImageProcessorImpl.cpp",
+      "../common/ota/OTAImageProcessorImpl.h",
+      "../common/ota/OTATlvProcessor.cpp",
+      "../common/ota/OTATlvProcessor.h",
     ]
 
     if (chip_enable_ota_firmware_processor == 1) {
